web-app/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
Click==7.0
Flask==1.1.1
idna==3.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
PyJWT==2.6.0
python-dotenv==0.10.3
pytz==2022.7.1
requests==2.28.2
twilio==7.17.0
urllib3==1.26.15
Werkzeug==0.16.0
pymongo